Loop flip
Fliped some loops to go backwards for optimastion

+++ src/ft_colorcode.c
@@ -6,13 +6,13 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/03/18 16:13:28 by pholster       #+#    #+#                */
-/*   Updated: 2019/07/21 22:45:18 by pholster      ########   odam.nl         */
+/*   Updated: 2019/08/10 11:18:36 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
 
-int		ft_colorcode(const char *str)
+ssize_t		ft_colorcode(const char *str)
 {
 	return (ft_colorncode(str, 16));
 }
